Fix for long datatype (removed in py3)

+++ googleapiclient/discovery.py
@@ -65,6 +65,7 @@ from googleapiclient.model import JsonModel
 from googleapiclient.model import MediaModel
 from googleapiclient.model import RawModel
 from googleapiclient.schema import Schemas
+from googleapiclient.compat import long
 from oauth2client.client import GoogleCredentials
 from oauth2client.util import _add_query_parameter
 from oauth2client.util import positional
@@ -330,7 +331,7 @@ def _media_size_to_long(maxSize):
       The size as an integer value.
     """
     if len(maxSize) < 2:
-        return 0L
+        return long(0)
     units = maxSize[-2:].upper()
     bit_shift = _MEDIA_SIZE_BIT_SHIFTS.get(units)
     if bit_shift is not None:
